requirements.txt
augraphy==8.2.6
Brotli==1.1.0
ConfigParser==7.2.0
cryptography==44.0.2
Cython==3.0.12
dl==0.1.0
docutils==0.21.2
fitz==0.0.1.dev2
fpdf2==2.8.2
gTTS==2.5.4
HTMLParser==0.0.2
importlib_metadata==6.11.0
keras==2.10.0
keyring==25.6.0
lockfile==0.12.2
lxml==5.3.1
matplotlib==3.10.1
numpy==1.24.3
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.11.0.86
ordereddict==1.1
paddleocr==2.7.0.0
paddleocr.egg==info
Pillow==11.1.0
protobuf==3.19.6
PyMuPDF==1.20.2
pyOpenSSL==25.0.0
PyPDF2==3.0.1
pywin32==310
scikit_learn==1.3.0
streamlit==1.22.0
tensorflow==2.10.0
thread==2.0.5
toml==0.10.2
tornado==6.4.2
xmlrpclib==1.0.1
zipp==3.21.0
